refactor : move controller classes into upper hierarchy package
