+++ src/main/java/com/ironknee/Knee2KneelSpring/controller/StageController.java
@@ -1,4 +1,4 @@
-package com.ironknee.Knee2KneelSpring.controller.stage;
+package com.ironknee.Knee2KneelSpring.controller;
 
 import org.springframework.web.bind.annotation.RestController;
